[#355] cleanup (moved some classes to search package)
