+++ src/main/java/de/bonndan/nivio/input/DiffProcessor.java
@@ -3,6 +3,7 @@ package de.bonndan.nivio.input;
 import de.bonndan.nivio.input.dto.ItemDescription;
 import de.bonndan.nivio.input.dto.LandscapeDescription;
 import de.bonndan.nivio.model.*;
+import de.bonndan.nivio.search.ItemMatcher;
 
 import java.util.*;
 import java.util.stream.Collectors;
