+++ src/main/java/de/bonndan/nivio/input/ItemRelationProcessor.java
@@ -2,6 +2,7 @@ package de.bonndan.nivio.input;
 
 import de.bonndan.nivio.input.dto.LandscapeDescription;
 import de.bonndan.nivio.model.*;
+import de.bonndan.nivio.search.ItemMatcher;
 
 import java.util.Iterator;
