+++ src/test/java/de/bonndan/nivio/search/ItemMatcherTest.java
@@ -1,6 +1,8 @@
-package de.bonndan.nivio.model;
+package de.bonndan.nivio.search;
 
 import de.bonndan.nivio.input.dto.ItemDescription;
+import de.bonndan.nivio.model.FullyQualifiedIdentifier;
+import de.bonndan.nivio.search.ItemMatcher;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
